add roles Administrator, Moderator, Developer

+++ backend/src/main/java/ru/compito/taskmanager/controller/AttributeController.java
@@ -4,10 +4,15 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 import ru.compito.taskmanager.config.ServiceConstants;
 import ru.compito.taskmanager.entity.Attribute;
+import ru.compito.taskmanager.entity.CustomField;
 import ru.compito.taskmanager.service.AttributeService;
+import ru.compito.taskmanager.service.ContentRelatedRoleService;
+import ru.compito.taskmanager.service.CustomFieldService;
 
 import java.util.List;
 
@@ -17,6 +22,10 @@ public class AttributeController {
 
     @Autowired
     private AttributeService attributeService;
+    @Autowired
+    private CustomFieldService customFieldService;
+    @Autowired
+    private ContentRelatedRoleService contentRelatedRoleService;
 
     @GetMapping(value = "/", produces = MediaType.APPLICATION_JSON_VALUE)
     public @ResponseBody ResponseEntity<Object> getAll() {
@@ -33,13 +42,32 @@ public class AttributeController {
     @PutMapping(value = "/{attributeId}/",
             consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
     public @ResponseBody ResponseEntity<Attribute> update(@PathVariable Integer attributeId, @RequestBody Attribute updatedAttribute) {
-        Attribute attribute = attributeService.update(attributeId, updatedAttribute);
-        return new ResponseEntity<>(attribute, HttpStatus.OK);
+        CustomField customField = customFieldService.getByAttribute(attributeId);
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        Integer boardId = customField.getTask().getBoard().getId();
+        if(contentRelatedRoleService.isContentOwner(boardId,authentication) ||
+                contentRelatedRoleService.isContentAdministrator(boardId,authentication)||
+                contentRelatedRoleService.isContentModerator(boardId,authentication)) {
+            Attribute attribute = attributeService.update(attributeId, updatedAttribute);
+            return new ResponseEntity<>(attribute, HttpStatus.OK);
+        }else {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
     }
 
     @DeleteMapping("/{attributeId}/")
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void delete(@PathVariable Integer attributeId) {
-        attributeService.delete(attributeId);
+    public @ResponseBody ResponseEntity<?> delete(@PathVariable Integer attributeId) {
+        CustomField customField = customFieldService.getByAttribute(attributeId);
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        Integer boardId = customField.getTask().getBoard().getId();
+        if(contentRelatedRoleService.isContentOwner(boardId,authentication) ||
+                contentRelatedRoleService.isContentAdministrator(boardId,authentication)||
+                contentRelatedRoleService.isContentModerator(boardId,authentication)) {
+            attributeService.delete(attributeId);
+            return new ResponseEntity<>(HttpStatus.OK);
+        }else {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
     }
 }
