+++ backend/src/main/java/ru/compito/taskmanager/controller/TemplateController.java
@@ -4,9 +4,14 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 import ru.compito.taskmanager.config.ServiceConstants;
+import ru.compito.taskmanager.entity.Task;
 import ru.compito.taskmanager.entity.TaskTemplate;
+import ru.compito.taskmanager.service.ContentRelatedRoleService;
+import ru.compito.taskmanager.service.TaskService;
 import ru.compito.taskmanager.service.TaskTemplateService;
 
 import java.util.List;
@@ -17,6 +22,10 @@ public class TemplateController {
 
     @Autowired
     private TaskTemplateService taskTemplateService;
+    @Autowired
+    private ContentRelatedRoleService contentRelatedRoleService;
+    @Autowired
+    private TaskService taskService;
 
     @GetMapping(value = "/", produces = MediaType.APPLICATION_JSON_VALUE)
     public @ResponseBody ResponseEntity<Object> getAll() {
@@ -35,20 +44,40 @@ public class TemplateController {
     public @ResponseBody ResponseEntity<TaskTemplate> create(@RequestBody TaskTemplate taskTemplate) {
         TaskTemplate newTaskTemplate = taskTemplateService.save(taskTemplate);
         return new ResponseEntity<>(newTaskTemplate, HttpStatus.CREATED);
+
+
     }
 
     @PutMapping(value = "/{taskTemplateId}/",
             consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
     public @ResponseBody ResponseEntity<TaskTemplate> update(@PathVariable Integer taskTemplateId, @RequestBody TaskTemplate updatedTaskTemplate) {
-        TaskTemplate taskTemplate = taskTemplateService.update(taskTemplateId,updatedTaskTemplate);
-        return new ResponseEntity<>(taskTemplate, HttpStatus.OK);
-
+        Task task  = taskService.getByTaskTemplateId(taskTemplateId);
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        Integer boardId = task.getBoard().getId();
+        if(contentRelatedRoleService.isContentOwner(boardId,authentication) ||
+                contentRelatedRoleService.isContentAdministrator(boardId,authentication)||
+                contentRelatedRoleService.isContentModerator(boardId,authentication)) {
+            TaskTemplate taskTemplate = taskTemplateService.update(taskTemplateId,updatedTaskTemplate);
+            return new ResponseEntity<>(taskTemplate, HttpStatus.OK);
+        }else {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
     }
 
     @DeleteMapping("/{taskTemplateId}/")
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void delete(@PathVariable Integer taskTemplateId) {
-        taskTemplateService.delete(taskTemplateId);
+    public @ResponseBody ResponseEntity<?> delete(@PathVariable Integer taskTemplateId) {
+        Task task  = taskService.getByTaskTemplateId(taskTemplateId);
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        Integer boardId = task.getBoard().getId();
+        if(contentRelatedRoleService.isContentOwner(boardId,authentication) ||
+                contentRelatedRoleService.isContentAdministrator(boardId,authentication)||
+                contentRelatedRoleService.isContentModerator(boardId,authentication)) {
+            taskTemplateService.delete(taskTemplateId);
+            return new ResponseEntity<>(HttpStatus.OK);
+        }else {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
     }
